+++ pyexcel_io/plugins.py
@@ -22,21 +22,21 @@ ERROR_MESSAGE_FORMATTER = "one of these plugins for %s data in '%s': %s"
 UPGRADE_MESSAGE = "Please upgrade the plugin '%s' according to \
 plugin compactibility table."
 
+text_stream_types = []
+binary_stream_types = []
+
 
 class IOManager(PluginManager):
-    def __init__(self):
-        PluginManager.__init__(self, 'pyexcel io plugin')
-        self.registry = defaultdict(list)
-        self.text_stream_types = []
-        self.binary_stream_types = []
-        self.loaded_reader_registry = defaultdict(dict)
-        self.loaded_writer_registry = defaultdict(dict)
+    def __init__(self, plugin_type, known_list):
+        PluginManager.__init__(self, plugin_type)
+        self.loaded_registry = defaultdict(dict)
+        self.known_plugins = known_list
 
     def load_me_later(self, plugin_meta, module_name):
         PluginManager.load_me_later(self, plugin_meta, module_name)
         if not isinstance(plugin_meta, dict):
-            plugin = module_name.replace('_', '-')
-            raise exceptions.UpgradePlugin(UPGRADE_MESSAGE % plugin)
+            pypi_name = _get_me_pypi_package_name(module_name)
+            raise exceptions.UpgradePlugin(UPGRADE_MESSAGE % pypi_name)
         library_import_path = "%s.%s" % (module_name, plugin_meta['submodule'])
         for file_type in plugin_meta['file_types']:
             self.registry[file_type].append(
@@ -55,9 +55,10 @@ class IOManager(PluginManager):
     def register_a_plugin(self, cls):
         if hasattr(cls, 'file_types'):
             PluginManager.register_a_plugin(self, cls)
+            pypi_name = _get_me_pypi_package_name(cls.__module__)
             for file_type in cls.file_types:
                 self._register_a_plugin(
-                    cls.action, file_type, cls, cls.library)
+                    cls.action, file_type, cls, pypi_name)
                 manager.register_a_file_type(
                     file_type, cls.stream_type, None)
         else:
@@ -65,19 +66,11 @@ class IOManager(PluginManager):
                 "not register abstract interface %s" % cls.__name__)
 
     def _register_a_plugin(self, action, file_type, plugin, library):
-        registry = self.loaded_reader_registry
-        if action == 'write':
-            registry = self.loaded_writer_registry
-
-        registry[file_type][library] = plugin
+        self.loaded_registry[file_type][library] = plugin
 
     def get_a_plugin(self, action, file_type, library):
         __file_type = file_type.lower()
-        registry = self.loaded_reader_registry
-        known_plugins = ioutils.AVAILABLE_READERS
-        if action == 'write':
-            registry = self.loaded_writer_registry
-            known_plugins = ioutils.AVAILABLE_WRITERS
+        registry = self.loaded_registry
 
         self.load_me_now(__file_type)
 
@@ -94,7 +87,7 @@ class IOManager(PluginManager):
             handler = handler_cls()
             handler.set_type(__file_type)
             return handler
-        plugins = known_plugins.get(file_type, None)
+        plugins = self.known_plugins.get(file_type, None)
         if plugins:
             message = "Please install "
             if len(plugins) > 1:
@@ -107,22 +100,21 @@ class IOManager(PluginManager):
             raise exceptions.NoSupportingPluginFound(
                 "No suitable library found for %s" % file_type)
 
-    def get_all_reader_formats(self):
-        all_formats = set(list(self.loaded_reader_registry.keys()) +
-                          list(ioutils.AVAILABLE_READERS.keys()))
-        all_formats = all_formats.difference(set([constants.DB_SQL,
-                                                  constants.DB_DJANGO]))
-        return all_formats
-
-    def get_all_writer_formats(self):
-        all_formats = set(list(self.loaded_writer_registry.keys()) +
-                          list(ioutils.AVAILABLE_WRITERS.keys()))
+    def get_all_formats(self):
+        all_formats = set(list(self.loaded_registry.keys()) +
+                          list(self.known_plugins.keys()))
         all_formats = all_formats.difference(set([constants.DB_SQL,
                                                   constants.DB_DJANGO]))
         return all_formats
 
 
-iomanager = IOManager()
+def _get_me_pypi_package_name(module_name):
+    root_module_name = module_name.split('.')[0]
+    return root_module_name.replace('_', '-')
+
+
+readers = IOManager('pyexcel-io reader', ioutils.AVAILABLE_READERS)
+writers = IOManager('pyexcel-io writer', ioutils.AVAILABLE_WRITERS)
 
 
 def load_plugins(prefix, path, black_list, white_list):