+++ pyexcel_io/readers/__init__.py
@@ -1,0 +1,26 @@
+__pyexcel_io_plugins__ = [
+    {
+        'plugin_type': 'pyexcel-io reader',
+        'submodule': 'csvr',
+        'file_types': ['csv'],
+        'stream_type': 'text'
+    },
+    {
+        'plugin_type': 'pyexcel-io reader',
+        'submodule': 'tsv',
+        'file_types': ['tsv'],
+        'stream_type': 'text'
+    },
+    {
+        'plugin_type': 'pyexcel-io reader',
+        'submodule': 'csvz',
+        'file_types': ['csvz'],
+        'stream_type': 'binary'
+    },
+    {
+        'plugin_type': 'pyexcel-io reader',
+        'submodule': 'tsvz',
+        'file_types': ['csvz'],
+        'stream_type': 'binary'
+    }
+ ]