+++ pyexcel_io/writers/__init__.py
@@ -1,24 +1,24 @@
 __pyexcel_io_plugins__ = [
     {
-        'plugin_type': 'pyexcel io plugin',
-        'submodule': '_csv',
+        'plugin_type': 'pyexcel-io writer',
+        'submodule': 'csvw',
         'file_types': ['csv'],
         'stream_type': 'text'
     },
     {
-        'plugin_type': 'pyexcel io plugin',
+        'plugin_type': 'pyexcel-io writer',
         'submodule': 'tsv',
         'file_types': ['tsv'],
         'stream_type': 'text'
     },
     {
-        'plugin_type': 'pyexcel io plugin',
+        'plugin_type': 'pyexcel-io writer',
         'submodule': 'csvz',
         'file_types': ['csvz'],
         'stream_type': 'binary'
     },
     {
-        'plugin_type': 'pyexcel io plugin',
+        'plugin_type': 'pyexcel-io writer',
         'submodule': 'tsvz',
         'file_types': ['tsvz'],
         'stream_type': 'binary'