+++ tests/test_django_book.py
@@ -2,11 +2,17 @@ from nose.tools import raises, eq_
 from pyexcel_io import save_data
 from pyexcel_io._compact import OrderedDict
 from pyexcel_io.constants import DB_DJANGO
-from pyexcel_io.database.django import (
-    DjangoModelReader, DjangoModelWriter,
-    DjangoModelImporter, DjangoModelExporter,
-    DjangoModelImportAdapter, DjangoModelExportAdapter,
-    DjangoBookWriter, DjangoBookReader
+from pyexcel_io.database.importers.django import (
+    DjangoModelWriter,
+    DjangoModelImporter,
+    DjangoModelImportAdapter,
+    DjangoBookWriter
+)
+from pyexcel_io.database.exporters.django import (
+    DjangoModelReader,
+    DjangoModelExporter,
+    DjangoModelExportAdapter,
+    DjangoBookReader
 )